Created htt auth some api request added PUT request to films

+++ film_project/services/web/project/film_resources.py
@@ -1,11 +1,25 @@
 """Module responsible for various resources that responsible
 for GET, POST, DELETE request processing on Film, Director models"""
 
-from flask import jsonify, abort, request
+from flask import jsonify, abort, request, Blueprint
 from flask_restx import Resource
+from flask_login import current_user, login_required
+from flask_httpauth import HTTPBasicAuth
 
 
 from . import models
+
+
+films_blueprint = Blueprint('films_blueprint', __name__, template_folder='templates')
+auth = HTTPBasicAuth()
+
+
+@auth.verify_password
+def verify_password(username, password):
+    user = models.User.query.filter_by(username=username).first()
+    if user and user.password == password:
+        return True
+    return False
 
 
 class FilmResource(Resource):
@@ -18,17 +32,30 @@ class FilmResource(Resource):
             abort(404)
         return jsonify(film.serialize)
 
+    @auth.login_required
     def post(self):
-        """POST request creates new film in database"""
-        user_id = request.json["user_id"]
+        """POST request creates new film in database only for authorized users"""
+        user = models.User.query.filter_by(username=auth.current_user()).first()
+        director_name = request.json.get("director_name")
+        director_surname = request.json.get("director_surname")
+        if director_surname is not None:
+            director = models.Director.query.filter_by(name=director_name, surname=director_surname).first()
+            if director:
+                director_id = director.id
+            else:
+                new_director = models.Director(name=director_name, surname=director_surname)
+                models.db.session.add(new_director)
+                director_id = models.Director.query.filter_by(name=director_name, surname=director_surname).first().id
+                models.db.session.commit()
+        else:
+            director_id = None
+        user_id = user.id
         name = request.json["name"]
         release_date = request.json["release_date"]
         rating = request.json['rating']
         poster_link = request.json['poster_link']
         description = request.json.get('description', '')
-        director_id = request.json.get('director_id', None)
         res = models.Film.query.filter_by(name=name).first()
-        print(res)
         if rating > 10 or rating < 1:
             abort(400)
         if res:
@@ -40,12 +67,62 @@ class FilmResource(Resource):
         models.db.session.commit()
         return jsonify(new_film.serialize)
 
-    def delete(self, film_id):
-        """DELETE request deletes Film by id"""
+    @auth.login_required
+    def put(self, film_id):
+        """PUT request to change films rating, poster_link, name, director only by admin or user who added it"""
+        user = models.User.query.filter_by(username=auth.current_user()).first()
         film = models.Film.query.get_or_404(film_id)
-        models.db.session.delete(film)
-        models.db.session.commit()
-        return "", 204
+        rating = request.json.get('rating')
+        poster_link = request.json.get('poster_link')
+        name = request.json.get('name')
+        director_name = request.json.get("director_name")
+        director_surname = request.json.get("director_surname")
+        if rating > 10 or rating < 1:
+            abort(400)
+        if film.user_id == user.id or user.is_admin:
+            if rating:
+                models.db.session.query(models.Film).filter(models.Film.id == film.id). \
+                    update({models.Film.rating: rating})
+                models.db.session.commit()
+
+            if poster_link:
+                models.db.session.query(models.Film).filter(models.Film.id == film.id). \
+                    update({models.Film.poster_link: poster_link})
+                models.db.session.commit()
+
+            if name:
+                models.db.session.query(models.Film).filter(models.Film.id == film.id). \
+                    update({models.Film.name: name})
+                models.db.session.commit()
+
+            if director_name and director_surname:
+                director = models.Director.query.filter_by(name=director_name, surname=director_surname).first()
+                if director:
+                    models.db.session.query(models.Film).filter(models.Film.id == film.id). \
+                        update({models.Film.director_id: director.id})
+                    models.db.session.commit()
+                else:
+                    new_director = models.Director(name=director_name, surname=director_surname)
+                    models.db.session.add(new_director)
+                    director_id = models.Director.query.filter_by(name=director_name,
+                                                                  surname=director_surname).first().id
+                    models.db.session.query(models.Film).filter(models.Film.id == film.id). \
+                        update({models.Film.director_id: director_id})
+                    models.db.session.commit()
+
+        return jsonify(film.serialize)
+
+    @auth.login_required
+    def delete(self, film_id):
+        """DELETE request deletes Film by id only by user that added it or admin"""
+        user = models.User.query.filter_by(username=auth.current_user()).first()
+        film = models.Film.query.get_or_404(film_id)
+        if film.user_id == user.id or user.is_admin:
+            models.db.session.delete(film)
+            models.db.session.commit()
+            return "", 204
+        else:
+            abort(403)
 
 
 class FilmDirector(Resource):
@@ -59,6 +136,9 @@ class FilmDirector(Resource):
         films = []
         for film, director in film_directors:
             films.append(film)
+
+        if len(films) == 0:
+            abort(404)
 
         return jsonify({f"{director_id}": [i.serialize for i in films]})
 
@@ -114,13 +194,16 @@ class FilmsResource(Resource):
     def get(self):
         """GET request returns all Films in JSON"""
         films = models.Film.query.all()
+        for film in films:
+            if film.director_id is None:
+                film.director_id = 'unknown'
         return jsonify(list=[i.serialize for i in films])
 
 
 class DirectorResource(Resource):
     """Resource responsible for GET, POST and DELETE request with director"""
     def get(self, director_id):
-        """GET request returns Film in JSON format by id"""
+        """GET request returns director in JSON format by id"""
         director = models.Director.query.filter_by(id=director_id).first()
         if not director:
             abort(404)
@@ -135,11 +218,16 @@ class DirectorResource(Resource):
         models.db.session.commit()
         return jsonify(new_director.serialize)
 
+    @auth.login_required
     def delete(self, director_id):
-        """DELETE request deletes director by id"""
-        director = models.Director.query.get_or_404(director_id)
-        models.db.session.query(models.Film).filter(models.Film.director_id == director.id).\
-            update({models.Film.director_id: None})
-        models.db.session.delete(director)
-        models.db.session.commit()
-        return "", 204
+        """DELETE request deletes director by id only for admins"""
+        user = models.User.query.filter_by(username=auth.current_user()).first()
+        if user.is_admin:
+            director = models.Director.query.get_or_404(director_id)
+            models.db.session.query(models.Film).filter(models.Film.director_id == director.id).\
+                update({models.Film.director_id: None})
+            models.db.session.delete(director)
+            models.db.session.commit()
+            return "", 204
+        elif not user.is_admin:
+            abort(403)
